+++ services/device-registry-base/src/main/java/org/eclipse/hono/service/management/tenant/Tenant.java
@@ -73,11 +73,12 @@ public class Tenant {
     private ResourceLimits resourceLimits;
 
     @JsonProperty(RegistryManagementConstants.FIELD_TRACING)
-    @JsonInclude(Include.NON_EMPTY)
+    @JsonInclude(Include.NON_NULL)
     private TenantTracingConfig tracing;
 
     @JsonProperty(RegistryManagementConstants.FIELD_PAYLOAD_TRUSTED_CA)
-    private List<TrustedCertificateAuthority> trustedCertificateAuthorities;
+    @JsonInclude(Include.NON_EMPTY)
+    private List<TrustedCertificateAuthority> trustedCertificateAuthorities = List.of();
 
     /**
      * Creates a new Tenant instance.
@@ -108,7 +109,7 @@ public class Tenant {
         this.resourceLimits = other.resourceLimits;
         this.tracing = other.tracing;
         if (Objects.nonNull(other.trustedCertificateAuthorities)) {
-            this.trustedCertificateAuthorities = new ArrayList<>(other.trustedCertificateAuthorities);
+            this.trustedCertificateAuthorities = List.copyOf(other.trustedCertificateAuthorities);
         }
     }
 
@@ -367,7 +368,7 @@ public class Tenant {
      */
     public Tenant setTrustedCertificateAuthorities(final List<TrustedCertificateAuthority> trustedCertificateAuthorities) {
         if (trustedCertificateAuthorities != null) {
-            this.trustedCertificateAuthorities = Collections.unmodifiableList(trustedCertificateAuthorities);
+            this.trustedCertificateAuthorities = List.copyOf(trustedCertificateAuthorities);
         }
         return this;
     }